Share the rendered-text helpers between the tab suites
BibleTabTestSupport and ScheduleTabTestSupport each declared their own identical
renderedText/showsExactly/showsContainingText at the top level of the same
package. Each compiled while it was the only one present, so both PRs were green
in isolation and the clash only appeared once #78 landed on main: every call site
became an overload resolution ambiguity.

They describe no particular tab, so they move to TabRenderedText.kt and both
support files use that copy.
